Add WebSocket Endpoint for Expiration Events
A tiny bit of cleanup too :)

+++ src/main/java/io/arsha/api/config/services/CacheConfigurationService.java
@@ -16,7 +16,7 @@ public class CacheConfigurationService {
         return cacheProperties.getRedis();
     }
 
-    public Integer getTtl() {
+    public Long getTtl() {
         return cacheProperties.getTtl();
     }
